Part Updates
UV Bezel - The cutout for the switch has been made larger to accommodate the switch I changed over to.

MultiCart - My height measurement was .51mm off resulting in a very snug fit with the cart.  I reduced the height over by .51mm and added two more screw bosses in the corners of the cart to help with the cart to be easily pulled open.

Steering Controller -  Modified housing to fit an alternate potentiometer.